Additional reqs

+++ nianalysis/interfaces/utils.py
@@ -136,6 +136,9 @@ class MergeTuple(Merge):
         return outputs
 
 
+
+
+
 class Chain(IdentityInterface):
 
     def _list_outputs(self):
@@ -367,9 +370,12 @@ class UnzipDir(CommandLine):
 
 
 class CopyToDirInputSpec(TraitedSpec):
-    in_files = traits.List(File(exists=True), mandatory=True,
+    in_files = traits.List(traits.Either(File(exists=True),
+                                         Directory(exists=True)), mandatory=True,
                            desc='input dicom files')
     out_dir = File(genfile=True, desc='the output dicom file')
+    extension = traits.Str(desc='specify the extention for the copied file.',
+                           default='', usedefault=True)
 
 
 class CopyToDirOutputSpec(TraitedSpec):
@@ -378,7 +384,7 @@ class CopyToDirOutputSpec(TraitedSpec):
 
 class CopyToDir(BaseInterface):
     """
-    Copies a list of files into a directory
+    Copies a list of files of directories into a directory
     """
 
     input_spec = CopyToDirInputSpec
@@ -387,9 +393,20 @@ class CopyToDir(BaseInterface):
     def _run_interface(self, runtime):
         dirname = self._gen_outdirname()
         os.makedirs(dirname)
+        ext = self.inputs.extension
         for i, f in enumerate(self.inputs.in_files):
-            fname = os.path.join(dirname, str(i).zfill(4)) + '.dcm'
-            shutil.copy(f, fname)
+            if os.path.isdir(f):
+                out_name = f.split('/')[-1]
+                if ext:
+                    out_name = '{0}_{1}'.format(
+                        out_name, ext+str(i).zfill(3))
+                shutil.copytree(f, dirname+'/{}'.format(out_name))
+            elif os.path.isfile(f):
+                if ext == '.dcm':
+                    fname = os.path.join(dirname, str(i).zfill(4)) + ext
+                else:
+                    fname = dirname
+                shutil.copy(f, fname)
         return runtime
 
     def _list_outputs(self):
@@ -408,7 +425,7 @@ class CopyToDir(BaseInterface):
         if isdefined(self.inputs.out_dir):
             dpath = self.inputs.out_dir
         else:
-            dpath = os.path.join(os.getcwd(), 'dicom_dir')
+            dpath = os.path.join(os.getcwd(), 'store_dir')
         return dpath
